+++ src/main/kotlin/com/github/wiggglez/charles_schwab_api/data_objs/responses/OptionChainResponse.kt
@@ -54,7 +54,7 @@ internal data class OptionResponse(
     val theta: Double,
     val vega: Double,
     val rho: Double,
-    val openInterest: Int,
+    val openInterest: Long,
     val timeValue: Double,
     val theoreticalOptionValue: Double,
     val theoreticalVolatility: Double,
